==> eda/01_data_exploration_bears.ipynb <==
# Import dependencies
import pandas as pd
import numpy as np
import hvplot.pandas

# Extend width of the results
pd.set_option('max_colwidth', 400)

### Bear Attacks data retrieval

# Read the data into a Pandas DataFrame
bear_data_raw_df = pd.read_csv("Resources/source/bear_attacks.csv")
bear_data_raw_df.head()

# Get a brief summary of the shark_data_raw_df DataFrame
bear_data_raw_df.info()

# Brief summary of the shark_data_raw_df DataFrame
bear_data_raw_df.describe()

### Bear Data Cleaning

# Copy the raw data to a new DataFrame
bear_data_df = bear_data_raw_df.copy()
bear_data_df.head()

bear_data_df.info()

# Change headers to lowercase and replace spaces with underscores
bear_data_df.columns = bear_data_df.columns.str.lower().str.replace(" ", "_")
bear_data_df.head()

#### Clean Date

# Verify the original date values
print(bear_data_df.date.unique())

# Replace specific text patters in the 'Date' column (e.g. 'Reported ')
bear_data_df['date'] = bear_data_df['date'].str.replace('Reported ', '', regex=False)

# Convert 'Date' to datetime
bear_data_df['date'] = pd.to_datetime(bear_data_df['date'], errors='coerce', infer_datetime_format=True)

# Function to extract year, month, and day from datetime
def extract_date_components(date):
    if pd.notnull(date):
        return pd.Series([date.year, date.month, date.day])
    else:
        return pd.Series([np.nan, np.nan, np.nan])

# Apply the function to the 'Date' column and create new columns for year, month, and day
bear_data_df[['year', 'month', 'day']] = bear_data_df['date'].apply(extract_date_components)

# Display the cleaned DataFrame with Year, Month, and Day
print(bear_data_df[['date', 'year', 'month', 'day']].head())

# Verify the distinct values in the new columns
distinct_years = bear_data_df['year'].unique()
distinct_months = bear_data_df['month'].unique()
distinct_days = bear_data_df['day'].unique()

print(f"Distinct Years: {distinct_years}")
print(f"Distinct Months: {distinct_months}")
print(f"Distinct Days: {distinct_days}")

date_string = bear_data_df['month'].astype(str) + '-' + bear_data_df['day'].astype(str) + '-' + bear_data_df['year'].astype(str)

date_string = date_string.replace('nan-nan-nan', np.nan)
date_string = date_string.str.replace('.0', '')

print(date_string.unique)

# Convert date to datetime
bear_data_df['date'] = pd.to_datetime(date_string, errors='coerce', format='%m-%d-%Y')
bear_data_df['date'].unique()

#### Clean age

# Verify the original values
bear_data_df.age.unique()

# Function to clean and convert age values
def clean_age(value):
    # Remove leading/trailing whitespace
    value = str(value).strip()
    
    # Handle special cases
    if 'month' in value or '?' in value or 'unknown' in value.lower():
        return np.nan
    
    # Convert to numeric, setting errors='coerce' to handle non-numeric values
    try:
        # Convert value to numeric, forcing errors to NaN
        return pd.to_numeric(value, errors='coerce')
    except Exception as e:
        print(f"Error processing age: {value} - {e}")
        return np.nan

# Apply the function to the 'Age' column
bear_data_df['age'] = bear_data_df['age'].apply(clean_age)

# Display the cleaned DataFrame with the 'Age' column
print(bear_data_df[['age']].head())

# Verify the distinct values in the 'Age' column
bear_data_df['age'].unique()

#### Clean gender

# Verify the original values
bear_data_df['gender'].unique()

# Define mapping for gender
gender_mapping = {
    ' male': 'M',
    ' female': 'F',
    '': 'U',  
    np.nan: 'U'  # Handles NaN values
}
# Replace values in the 'Gender' column using the mapping
bear_data_df['gender'] = bear_data_df['gender'].replace(gender_mapping)

# Rename 'Gender' to 'Sex'
bear_data_df.rename(columns={'gender': 'sex'}, inplace=True)

# Verify the updated values
bear_data_df['sex'].unique()

### Limit years to match years in Shark Attacks

# Limit the years in bear data from 1900 to 2019
bear_data_df = bear_data_df[(bear_data_df['year'] >= 1900) & (bear_data_df['year'] <= 2019)]
bear_data_df['year'].sort_values(ascending=False).unique()


### Display a map of Bear Attacks

%%capture --no-display

# Configure the map plot
bear_attack = bear_data_df[["date", "latitude", "longitude"]]

# Convert date to string
bear_attack["date"] = bear_attack["date"].dt.strftime('%Y-%m-%d')

# Configure the map plot
bear_map = bear_attack.hvplot.points(
    x="longitude",
    y="latitude",
    geo=True,
    tiles="OSM",
    color = "date",
    frame_width = 700,
    frame_height = 500
)

# Display the map
bear_map

### Save Bear Attacks dataframe into CSV file in the Resources directory

bear_data_df.to_csv("Resources/bear_attacks_clean_data.csv", index=False)
