==> eda/02_data_merge.ipynb <==
# Import dependencies
import pandas as pd
import numpy as np
import json

# from config import GEOAPIFY_API
import requests
import pprint
import time

pd.set_option('max_colwidth', 400)

# Read Shark Attacks clean data
shark_attacks_df = pd.read_csv("Resources/shark_attacks_clean_data.csv")
shark_attacks_df.head()

# Read Bear Attacks clean data
bear_attacks_df = pd.read_csv("Resources/bear_attacks_clean_data.csv")  
bear_attacks_df.head()


# Read Shark Attacks location data
shark_unique_locations_df = pd.read_csv("Resources/shark_unique_locations.csv")
shark_unique_locations_df.head()

# For bear_attacks_df
bear_attacks_df['animal_type'] = 'bear'
# For shark_attacks_df
shark_attacks_df['animal_type'] = 'shark'

# Check columns
shark_attacks_df.columns

# Merge the DataFrames on the 'location', 'area' and 'country' columns
merged_shark_df = pd.merge(shark_attacks_df, shark_unique_locations_df, on=['location', 'area', 'country'], how='inner')
merged_shark_df.columns

# Renark_sharkme columns in all DataFrames
merged_shark_df.rename(columns={
    'bear attack': 'bear',
    'date': 'date',
    'location': 'location',
    'name': 'name',
    'sex': 'sex',
    'year': 'year',
    'type': 'attack_type',
    'country': 'country',
    'area': 'area',
    'activity': 'activity',
    'species': 'species',
    'investigator_or_source': 'investigator_or_source',
    'pdf': 'pdf',
    'href_formula': 'href_formula',
    'href': 'href',
    'count': 'count',
    'latitude': 'latitude',
    'longitude': 'longitude'
}, inplace=True)

bear_attacks_df.rename(columns={
    'date': 'date',
    'location': 'location',
    'details': 'details',
    'bear': 'species',
    'latitude': 'latitude',
    'longitude': 'longitude',
    'name': 'name',
    'age': 'age',
    'sex': 'sex',
    'year': 'year',
    'month': 'month',
    'day': 'day'
}, inplace=True)



# Validate column names
print(merged_shark_df.columns)
print(bear_attacks_df.columns)

# Reset index for each DataFrame before concatenation
merged_shark_df = merged_shark_df.reset_index(drop=True)
bear_attacks_df = bear_attacks_df.reset_index(drop=True)

# Now concatenate the DataFrames
combined_df = pd.concat([merged_shark_df, bear_attacks_df], ignore_index=True, sort=False)


# Display the first few rows of the combined DataFrame
print(combined_df.head())

print(combined_df.tail())

# Select specific columns to view
selected_columns_df = combined_df[['date','sex','age']]


print(combined_df.describe())

# Export to CSV
combined_df.to_csv("Resources/combined_attacks.csv", index=False)
