==> eda/03_combined_attacks_to_sqlite.ipynb <==
# data science
import pandas as pd

# Python SQL toolkit and Object Relational Mapper
import sqlalchemy
from sqlalchemy import create_engine, inspect

df = pd.read_csv("Resources/combined_attacks.csv")
df.head()

### Store data in SQLite for EDA and Web App

# Function to store the SQLite database based on the location passed as a parameter
def store_db(directory):
    # Create the connection to the SQLite database
    engine = create_engine(f'sqlite:///{directory}/combined_attacks.sqlite')
    # Store the data in the SQLite database
    df.to_sql('combined_attacks', engine, if_exists='replace', index=False)

    # Return the INSPECT of the database to validate data being saved
    inspector_gadget = inspect(engine)

    # Close the engine connection
    engine.dispose()

    return inspector_gadget

# Store the database in the Resources folder
inspector_db = store_db("Resources")

# Collect the names of tables within the database
tables = inspector_db.get_table_names()

# print metadata for each table
for table in tables:
    print(table)
    print("-----------")
    
    # get columns
    columns = inspector_db.get_columns(table)
    for column in columns:
        print(column["name"], column["type"])

    print()

# Store the database in the Resources folder
inspector_db = store_db("../app/Resources")

# Collect the names of tables within the database
tables = inspector_db.get_table_names()

# print metadata for each table
for table in tables:
    print(table)
    print("-----------")
    
    # get columns
    columns = inspector_db.get_columns(table)
    for column in columns:
        print(column["name"], column["type"])

    print()
